move template implement code to its header file
